Implement delete event
